change id data type from Long to Integer;

+++ src/main/java/com/innowise/repo/AttachmentRepository.java
@@ -5,11 +5,11 @@ import com.innowise.domain.Attachment;
 import java.util.List;
 
 public interface AttachmentRepository {
-    Attachment findById(Long id);
+    Attachment findById(Integer id);
     List<Attachment> findAll();
 
     Attachment save(Attachment attachment);
     Attachment update(Attachment attachment);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
