+++ src/main/java/com/innowise/repo/CategoryRepository.java
@@ -5,11 +5,11 @@ import com.innowise.domain.Category;
 import java.util.List;
 
 public interface CategoryRepository {
-    Category findById(Long id);
+    Category findById(Integer id);
     List<Category> findAll();
 
     Category save(Category category);
     Category update(Category category);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
