+++ src/main/java/com/innowise/repo/CommentRepository.java
@@ -5,11 +5,11 @@ import com.innowise.domain.Comment;
 import java.util.List;
 
 public interface CommentRepository {
-    Comment findById(Long id);
+    Comment findById(Integer id);
     List<Comment> findAll();
 
     Comment save(Comment comment);
     Comment update(Comment comment);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
