+++ src/main/java/com/innowise/repo/HistoryRepository.java
@@ -5,11 +5,11 @@ import com.innowise.domain.History;
 import java.util.List;
 
 public interface HistoryRepository {
-    History findById(Long id);
+    History findById(Integer id);
     List<History> findAll();
 
     History save(History history);
     History update(History history);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
