+++ src/main/java/com/innowise/repo/UserRepository.java
@@ -5,11 +5,11 @@ import com.innowise.domain.User;
 import java.util.List;
 
 public interface UserRepository {
-    User findById(Long id);
+    User findById(Integer id);
     List<User> findAll();
 
     User save(User user);
     User update(User user);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
