+++ src/main/java/com/innowise/repo/impl/AttachmentRepositoryImpl.java
@@ -15,7 +15,7 @@ public class AttachmentRepositoryImpl implements AttachmentRepository {
     private Session session;
 
     @Override
-    public Attachment findById(Long id) {
+    public Attachment findById(Integer id) {
         return session.find(Attachment.class, id);
     }
 
@@ -35,7 +35,7 @@ public class AttachmentRepositoryImpl implements AttachmentRepository {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Integer id) {
         Optional<Attachment> attachmentToDelete = Optional.of(session.find(Attachment.class, id));
         if (attachmentToDelete.isEmpty()) {
             //todo throw exception
